Refactoring to RLCModel
