+++ packages/autorest.typescript/src/restLevelClient/generateTopLevelIndexFile.ts
@@ -5,41 +5,30 @@ import { Project } from "ts-morph";
 import * as path from "path";
 import { getAutorestOptions } from "../autorestSession";
 import { CodeModel } from "@autorest/codemodel";
-import { NameType, normalizeName } from "../utils/nameUtils";
-
-const batchOutputFolder: [string, string, string][] = [];
+import { transform } from "./transforms/transform";
+import { buildTopLevelIndex } from "@azure-tools/rlc-codegen";
 
 export function generateTopLevelIndexFile(model: CodeModel, project: Project) {
-  const { multiClient, batch, srcPath } = getAutorestOptions();
-  if (srcPath) {
-    const clientName = model.language.default.name;
-    const moduleName = normalizeName(clientName, NameType.File);
-    const relativePath = srcPath.replace("/src", "");
-    batchOutputFolder.push([relativePath, clientName, moduleName]);
-  }
+  const { srcPath } = getAutorestOptions();
+  const importedParameters = new Set<string>();
+  const importedResponses = new Set<string>();
+  const clientImports = new Set<string>();
+  const rlcModels = transform(model, {
+    importedParameters,
+    importedResponses,
+    clientImports
+  });
 
-  if (
-    multiClient &&
-    batch &&
-    batch.length > 1 &&
-    batchOutputFolder.length === batch.length
-  ) {
-    const { srcPath } = getAutorestOptions();
+  const preparedContent = buildTopLevelIndex(rlcModels);
+  if (preparedContent) {
     const fileDirectory = path.join(srcPath as string, "../../");
-    const file = project.createSourceFile("/src/index.ts", undefined, {
-      overwrite: true
-    });
-    file.moveToDirectory(fileDirectory);
-    const allModules: string[] = [];
-    batchOutputFolder.forEach(item => {
-      file.addImportDeclaration({
-        namespaceImport: item[1],
-        moduleSpecifier: `${item[0]}`
-      });
-      allModules.push(item[1]);
-    });
-    file.addExportDeclaration({
-      namedExports: [...allModules]
-    });
+    const indexFile = project.createSourceFile(
+      "/src/index.ts",
+      preparedContent.content,
+      {
+        overwrite: true
+      }
+    );
+    indexFile.moveToDirectory(fileDirectory);
   }
 }
