+++ packages/rlc-codegen/src/index.ts
@@ -4,10 +4,11 @@
 export * from "./interfaces.js";
 export * from "./generateClientDefinitions.js";
 export * from "./generateSchemaType.js";
-export * from './generateClient.js';
+export * from "./generateClient.js";
 export * from "./helpers/nameConstructors.js";
 export * from "./buildResponseTypes.js";
 export * from "./helpers/shortcutMethods.js";
 export * from "./helpers/nameUtils.js";
 export * from "./buildParameterTypes.js";
 export * from "./buildIsUnexpectedHelper.js";
+export * from "./buildTopLevelIndexFile.js";
